Split clocktower roles.ts into grouped modules
Extract the oversized roles.ts (482 lines, over the 400-line cap) into a
roles/ directory: _types.ts (enums + definition interface),
_distribution.ts (player-count distribution + defaultRoleCount), and one
file per character type (townsfolk, outsiders, minions, demons). index.ts
re-assembles CLOCKTOWER_ROLES and re-exports all public symbols, so every
importer keeps working unchanged. Pure refactor, no behavior change.
